grt: adding static_cast long for multiplication

+++ src/grt/src/fastroute/src/maze3D.cpp
@@ -860,11 +860,11 @@ void FastRouteCore::mazeRouteMSMDOrder3D(int expand,
   static multi_array<parent3D, 3> pr_3D_(
       boost::extents[num_layers_][y_grid_][x_grid_]);
 
-  long total_size = num_layers_ * y_range_ * x_range_;
+  long total_size = static_cast<long>(num_layers_) * y_range_ * x_range_;
   static std::vector<bool> pop_heap2_3D(total_size, false);
 
   // allocate memory for priority queue
-  total_size = y_grid_ * x_grid_ * num_layers_;
+  total_size = static_cast<long>(y_grid_) * x_grid_ * num_layers_;
   static std::vector<int*> src_heap_3D(total_size);
   static std::vector<int*> dest_heap_3D(total_size);
